frontend/requirements.txt
streamlit>=1.24.0
requests>=2.28.2
pandas>=1.5.3
python-dateutil>=2.8.2